Add sleep settings
Add a settings page for enabling/disabling actions to be done when
sleep is turned on or off.

Added Options:
- [x] Allow Always on Display
- [x] Allow Chimes
- [x] Allow Notifications
- [x] Disable BLE

Replace Allow AOD with Enable AOD

Leaving it as "Allow AOD" assumes the user wants it the same
during the day as during sleep. In my opinion the user would
either want AOD on during sleep or they wouldn't.