Timer Revamp
Add an hour option to the timer application

Add Timer's Time Remaining to the StatusIcons bar

When a timer is active the time remaining will be displayed in the
StatusIcons bar along with an hour glass symbol, and will be set to
hidden when the timer's off.

Add hour support to timer in StatusIcons.cpp

timer: Add ringing and counter

The timer app issues a short buzz once and then disappears. There is no
trace left that the timer finished or how long ago. This change makes
the motor start ringing and presents a timer counter.

The timer stops buzzing after 10 seconds, and finally resets after
1 minute.

Keep screen on during timer buzzing

This prevents the motorController from buzzing infinitely while the
watch is sleeping.